+++ src/commands/expressions/expression.ts
@@ -8,11 +8,8 @@ import type { AttributeValues } from "../attributes/values.js";
 
 // NOTES:
 // A string literal in an expression is *always* a reference to an attribute path, and seems to be always valid syntax.
-// For example, `beginws_with(toto, tata)` is valid and checks that the attribute value toto begins with the content of the attribute tata.
+// For example, `begins_with(toto, tata)` is valid and checks that the attribute value toto begins with the content of the attribute tata.
 // Obviously, this particular case is not useful, but there could be edge cases where this type of function call is, so we should support it.
-// We could always require the user to be explicit about it. For example, syntax like `beginsWith("toto", "tata")` could be rejected and the
-// user could have to provide one of: `beginsWith("toto", attribute("tata"))` or `beginsWith("toto", value("tata"))`. The call
-// can also entirely be made with attribute values, as such: `begins_with(:val1, :val2)`.
 
 // TODO: rename for serializer, and the method to be serialize.
 export type Stringifier = (params: {
@@ -33,21 +30,33 @@ export class Expression {
     this.stringifier = stringify;
   }
 
-  and(rhs: Expression): Expression {
+  /**
+   * Returns an expression that combines this one with the provided expression using the `AND` operator.
+   *
+   * @param other - The other expression to combine with this one.
+   * @returns An {@link Expression} that is true only if both expressions are true.
+   */
+  and(other: Expression): Expression {
     return new Expression({
       stringify: ({ attributeNames, attributeValues }) => {
         const left = this.stringify({ attributeNames, attributeValues });
-        const right = rhs.stringify({ attributeNames, attributeValues });
+        const right = other.stringify({ attributeNames, attributeValues });
         return `(${left} AND ${right})`;
       },
     });
   }
 
-  or(rhs: Expression): Expression {
+  /**
+   * Returns an expression that combines this one with the provided expression using the `OR` operator.
+   *
+   * @param other - The other expression to combine with this one.
+   * @returns An {@link Expression} that is true if any of the expressions is true.
+   */
+  or(other: Expression): Expression {
     return new Expression({
       stringify: ({ attributeNames, attributeValues }) => {
         const left = this.stringify({ attributeNames, attributeValues });
-        const right = rhs.stringify({ attributeNames, attributeValues });
+        const right = other.stringify({ attributeNames, attributeValues });
         return `(${left} OR ${right})`;
       },
     });
@@ -65,6 +74,12 @@ export function expression(params: ExpressionParams): Expression {
   return new Expression(params);
 }
 
+/**
+ * Negates the provided expression using the `NOT` operator.
+ *
+ * @param expression - The expression to negate.
+ * @returns An {@link Expression} that corresponds to the negation the provided expression.
+ */
 export function not(expression: Expression): Expression {
   return new Expression({
     stringify: ({ attributeNames, attributeValues }) => {
@@ -80,7 +95,25 @@ export abstract class Operand {
   beginsWith(rhs: Operand): Expression {
     return expression({
       stringify: ({ attributeNames, attributeValues }) => {
-        return `begins_with(${this.register({ attributeNames, attributeValues })}, ${rhs.register({ attributeNames, attributeValues })})`;
+        return `begins_with(${this.substitute({ attributeNames, attributeValues })}, ${rhs.substitute({ attributeNames, attributeValues })})`;
+      },
+    });
+  }
+
+  /**
+   * Returns an expression that checks if this operand is between the two provided ones using the `BETWEEN` operator.
+   *
+   * Both bounds are inclusive, meaning that the returned expression corresponds to `lower <= this <= upper`.
+   *
+   * @param lower - The lower inclusive bound of the range.
+   * @param upper - The upper inclusive bound of the range.
+   *
+   * @returns An {@link Expression} that is true if this operand is between the two provided bounds.
+   */
+  between(lower: Operand, upper: Operand): Expression {
+    return expression({
+      stringify: ({ attributeNames, attributeValues }) => {
+        return `${this.substitute({ attributeNames, attributeValues })} BETWEEN ${lower.substitute({ attributeNames, attributeValues })} AND ${upper.substitute({ attributeNames, attributeValues })}`;
       },
     });
   }
@@ -89,7 +122,7 @@ export abstract class Operand {
   contains(rhs: Operand): Expression {
     return expression({
       stringify: ({ attributeNames, attributeValues }) => {
-        return `contains(${this.register({ attributeNames, attributeValues })}, ${rhs.register({ attributeNames, attributeValues })})`;
+        return `contains(${this.substitute({ attributeNames, attributeValues })}, ${rhs.substitute({ attributeNames, attributeValues })})`;
       },
     });
   }
@@ -106,7 +139,7 @@ export abstract class Operand {
   equals(rhs: Operand): Expression {
     return expression({
       stringify: ({ attributeNames, attributeValues }) => {
-        return `${this.register({ attributeNames, attributeValues })} = ${rhs.register({ attributeNames, attributeValues })}`;
+        return `${this.substitute({ attributeNames, attributeValues })} = ${rhs.substitute({ attributeNames, attributeValues })}`;
       },
     });
   }
@@ -117,7 +150,7 @@ export abstract class Operand {
   eq = this.equals.bind(this);
 
   /**
-   * Tests that the left operand is greater than the right operand.
+   * Tests that this operand is greater than the right operand.
    *
    * This uses the `>` operator.
    *
@@ -128,13 +161,13 @@ export abstract class Operand {
   greaterThan(rhs: Operand): Expression {
     return expression({
       stringify: ({ attributeNames, attributeValues }) => {
-        return `${this.register({ attributeNames, attributeValues })} > ${rhs.register({ attributeNames, attributeValues })}`;
+        return `${this.substitute({ attributeNames, attributeValues })} > ${rhs.substitute({ attributeNames, attributeValues })}`;
       },
     });
   }
 
   /**
-   * Tests that the left operand is greater than or equals to the right operand.
+   * Tests that this operand is greater than or equals to the right operand.
    *
    * This uses the `>=` operator.
    *
@@ -145,7 +178,7 @@ export abstract class Operand {
   greaterThanOrEquals(rhs: Operand): Expression {
     return expression({
       stringify: ({ attributeNames, attributeValues }) => {
-        return `${this.register({ attributeNames, attributeValues })} >= ${rhs.register({ attributeNames, attributeValues })}`;
+        return `${this.substitute({ attributeNames, attributeValues })} >= ${rhs.substitute({ attributeNames, attributeValues })}`;
       },
     });
   }
@@ -161,7 +194,39 @@ export abstract class Operand {
   gte = this.greaterThanOrEquals.bind(this);
 
   /**
-   * Tests that the left operand is lower than the right operand.
+   * Tests that this operand is contained within the provided list.
+   *
+   * This uses the `IN` operator.
+   *
+   * @param operands - The list of operands to check against. This function throws if the list is
+   * empty or contains more than 100 operands.
+   *
+   * @returns The corresponding {@link Expression} that is true if this operand is contained within the provided list.
+   */
+  in(...operands: Operand[]): Expression {
+    if (operands.length === 0) {
+      throw new Error("the IN operator requires at least one operand.");
+    }
+    if (operands.length > 100) {
+      throw new Error(
+        `up to 100 operands are support for the IN operator, got ${operands.length}`,
+      );
+    }
+
+    return expression({
+      stringify: ({ attributeNames, attributeValues }) => {
+        const operandsString = operands
+          .map((operand) =>
+            operand.substitute({ attributeNames, attributeValues }),
+          )
+          .join(",");
+        return `${this.substitute({ attributeNames, attributeValues })} IN (${operandsString})`;
+      },
+    });
+  }
+
+  /**
+   * Tests that this operand is lower than the right operand.
    *
    * This uses the `<` operator.
    *
@@ -172,13 +237,13 @@ export abstract class Operand {
   lowerThan(rhs: Operand): Expression {
     return expression({
       stringify: ({ attributeNames, attributeValues }) => {
-        return `${this.register({ attributeNames, attributeValues })} < ${rhs.register({ attributeNames, attributeValues })}`;
+        return `${this.substitute({ attributeNames, attributeValues })} < ${rhs.substitute({ attributeNames, attributeValues })}`;
       },
     });
   }
 
   /**
-   * Tests that the left operand is lower than or equals to the right operand.
+   * Tests that this operand is lower than or equals to the right operand.
    *
    * This uses the `<=` operator.
    *
@@ -189,7 +254,7 @@ export abstract class Operand {
   lowerThanOrEquals(rhs: Operand): Expression {
     return expression({
       stringify: ({ attributeNames, attributeValues }) => {
-        return `${this.register({ attributeNames, attributeValues })} <= ${rhs.register({ attributeNames, attributeValues })}`;
+        return `${this.substitute({ attributeNames, attributeValues })} <= ${rhs.substitute({ attributeNames, attributeValues })}`;
       },
     });
   }
@@ -216,7 +281,7 @@ export abstract class Operand {
   notEquals(rhs: Operand): Expression {
     return expression({
       stringify: ({ attributeNames, attributeValues }) => {
-        return `${this.register({ attributeNames, attributeValues })} <> ${rhs.register({ attributeNames, attributeValues })}`;
+        return `${this.substitute({ attributeNames, attributeValues })} <> ${rhs.substitute({ attributeNames, attributeValues })}`;
       },
     });
   }
@@ -231,7 +296,7 @@ export abstract class Operand {
     return new SizeOperand(this);
   }
 
-  abstract register(params: {
+  abstract substitute(params: {
     attributeNames: AttributeNames;
     attributeValues: AttributeValues;
   }): string;
@@ -245,11 +310,29 @@ export class ExpressionAttribute extends Operand {
     this.path = path;
   }
 
+  /**
+   * @returns An {@link Expression} that returns true if the provided attribute path exists.
+   */
   // NOTE: the left hand side of this expression can only be a literal value (tested)
   exists(): Expression {
     return expression({
       stringify: ({ attributeNames, attributeValues }) => {
-        return `attribute_exists(${this.register({ attributeNames, attributeValues })})`;
+        return `attribute_exists(${this.substitute({ attributeNames, attributeValues })})`;
+      },
+    });
+  }
+
+  /**
+   * @param type - The type to check against.
+   * @returns An {@link Expression} that returns true if there exists an attribute at
+   * the provided path of the given type.
+   */
+  // NOTE: the left hand side of this expression *can be* an attribute value pointing to a valid path.
+  // NOTE: the right hand side of this expression *must be* an expression attribute (not a literal).
+  isType(type: ExpressionValue<AttributeType>): Expression {
+    return expression({
+      stringify: ({ attributeNames, attributeValues }) => {
+        return `attribute_type(${this.substitute({ attributeNames, attributeValues })}, ${type.substitute({ attributeNames, attributeValues })})`;
       },
     });
   }
@@ -258,22 +341,12 @@ export class ExpressionAttribute extends Operand {
   notExists(): Expression {
     return expression({
       stringify: ({ attributeNames, attributeValues }) => {
-        return `attribute_not_exists(${this.register({ attributeNames, attributeValues })})`;
+        return `attribute_not_exists(${this.substitute({ attributeNames, attributeValues })})`;
       },
     });
   }
 
-  // NOTE: the left hand side of this expression *can be* an attribute value pointing to a valid path.
-  // NOTE: the right hand side of this expression *must be* an expression attribute (not a literal).
-  type(type: AttributeType): Expression {
-    return expression({
-      stringify: ({ attributeNames, attributeValues }) => {
-        return `attribute_type(${this.register({ attributeNames, attributeValues })}, ${attributeValues.reference(type)})`;
-      },
-    });
-  }
-
-  register(params: {
+  substitute(params: {
     attributeNames: AttributeNames;
     attributeValues: AttributeValues;
   }): string {
@@ -286,15 +359,15 @@ export function attribute(path: AttributePath): ExpressionAttribute {
   return new ExpressionAttribute(path);
 }
 
-export class ExpressionValue extends Operand {
-  private readonly value: AttributeValue;
+export class ExpressionValue<T extends AttributeValue> extends Operand {
+  private readonly value: T;
 
-  constructor(value: AttributeValue) {
+  constructor(value: T) {
     super();
     this.value = value;
   }
 
-  register(params: {
+  substitute(params: {
     attributeNames: AttributeNames;
     attributeValues: AttributeValues;
   }): string {
@@ -303,7 +376,9 @@ export class ExpressionValue extends Operand {
   }
 }
 
-export function value(value: AttributeValue): ExpressionValue {
+export function value<T extends AttributeValue = AttributeValue>(
+  value: AttributeValue,
+): ExpressionValue<T> {
   return new ExpressionValue(value);
 }
 
@@ -315,10 +390,10 @@ class SizeOperand extends Operand {
     this.inner = operand;
   }
 
-  register(params: {
+  substitute(params: {
     attributeNames: AttributeNames;
     attributeValues: AttributeValues;
   }): string {
-    return `size(${this.inner.register(params)})`;
+    return `size(${this.inner.substitute(params)})`;
   }
 }